+++ crawler/requirements.txt
@@ -1,18 +1,11 @@
 # Selenium 관련
 selenium==4.16.0
-webdriver-manager==4.0.1
 
 # AWS SDK
 boto3==1.34.14
 
-# HTTP 클라이언트
-httpx==0.26.0
-
 # 환경변수 관리
 python-dotenv==1.0.0
 
-# 스케줄링 (나중에 사용)
-apscheduler==3.10.4
-
-# 유틸리티
+# HTTP 클라이언트
 requests==2.31.0
